refactor(wizards-of-react): remove double extension in case it fools bundlers
